доделать чуть

+++ WpfApp1/Models/tickets.cs
@@ -16,7 +16,7 @@ namespace WpfApp1.Models
     {
         public long id { get; set; }
         public long flight_id { get; set; }
-        public long seat_number { get; set; }
+        public int seat_number { get; set; }
         public long user_id { get; set; }
     
         public virtual flights flights { get; set; }
